BOARD-994: Adding some new code to the ExceptionRegister to support new incoming features
Adding some new features to the exceptions register. it will be possible at the beginning of a function to call GetRegisteredExceptionCountForCurrentThread() and record the value, and once all the try...catch() statement is done, do something like "if (unStartingExceptionCount < ::GetRegisteredExceptionCountForCurrentThread)" to determine if an exception was thrown within the current function, and then handle any cleanup code in ONE place, instead of handling it in all of the different catch statements (i.e. there will be three catch statements all the time, so it's a pain in the ass to duplicate the cleanup code in all of the catch blocks over and over again)