BOARD-1178
Modified the InitializerProcess to write the incoming dataset to file, pass the dataset as a filename back to the RootOfTrustCore. Modified the RootOfTrustNode to get the dataset filename instead of the actual dataset. Modified the DataConnector process (class) to now use the Dataset class, and read the dataset from file. Create the BinaryWriter class. Made some global changes to use one single Base64 (there was a conflict in RestFrameworkRuntimeData.cpp). So the Base64Encode() and Base64Decode() were removed from Utils.cpp and Utils.h