New built tool
The build tool has been augmented. Additional command line options are now available. The built tool also publishes header files and include files as well as part of the build step. A series of build.node files and the one build.root file have been updated; Currently, the 'demo' project which used to test the BuildTool is the DatasetViewer. So someone can go to the DatasetViewer folder and do a "BuildTool --this"

Currently, "BuiltTool --all" will not work, and that's because makefiles everywhere have to be reworked.

The BuildTool itself is actually the only project that cannot be built by the BuiltTool. One must call 'make all' independently. The compiled BuiltTool is now in the /BuildTools folder