BOARD-1253
Most of the bugs reported in this are not actual bugs, the tool is behaving properly. If there are 5 cells in the header, then any row without precisely 5 cells will be rejected. In most of the 'failed test' examples give, the header has 5 cells, but all the rows have 6 cells. The tools is supposed to reject all of the rows, that is the expected behavior.

The only real bug is the one with the '|' separator. It turns out that C# doesn't think that '|' is punctuation at all. Had to add code on line 101 to handle this weird exception.

Also added a new function which ensures that the same specification cannot be specified for more than one of the input fields.