Quick change to support datasets in the DatabaseInitializationTool
So this is actually a tiny minor change for now, but the real change involves a complete re-write of the DatabaseTools (see new folder DatabaseInitializationTool). But there's more. The ApiFunctionHelpers.cpp and .h should be made into their own static library and the singular big huge file should be cut up into a bunch of smaller files. We should also create a DigitalContract.h and DigitalContract.cpp class. Moreover, API calls into the SAIL Platform Services and SAIL Data services involving datasets and digitalcontract should use only Base64 versions of a serialized buffer since those are going to eventually be digitally signed. Right now, each element of a StructuredBuffer is put into a JSON, but we cannot have that in the future if we want to ensure that the digital signature is verifiable. The entire serialized version of the StructuredBuffer must be kept as-is. So changes to the back-end API's will be required.