BOARD-1554
So this was an interesting issue with DataGridView object in C# which, by default, is unable to host more than 300 columns. Has to add a special 'ColumnAdded' event to ensure that the fillweight of the column was set to 1 (i.e. max sum of all column fill weights can never exceed 65,535). Now the application is able to host 65,535 columns (but not more).