BOARD-1566 BOARD-1567 BOARD-1568 BOARD-1569 BOARD-1575
So wow, this was a really interesting bug. It turns out that the exact same copy of the ParseJsonString() function within the JsonParser.cpp (between Linux and Windows) have DIFFERENT behavior. It turns out that in Linux, if you do a strSomeString.push_back('/'), it will NOT insert the character.... But it does it in Windows without issue.

Anyway, changes we made to JsonParser.cpp to fix that bug, and some changes were made to the Orchestrator to better handle Session information. The way the original orchestrator did it could be improved. I've removed the EosbRotationManager.h and .cpp and replaced them with a SessionManager.h and .cpp. Also added a useful function in BinaryFileReader.cpp and .h to read a file all at once.