Fixing issues relating to this pull request
Some of the tests are now 'failing' because the behavior of GetJsonForStructuredBufferMap(), which is called by GetSafeFunctions(), GetDatasets(), GetDigitalContracts(), and GetTables() has changed. Instead of returning an empty string, it is returning a value empty JSON blob (i.e. "{ }"). This check in fixes this, along with a bug in the JsonParser.cpp