BOARD-407
A new DatasetInitializationTool has been created, and it will replace the existing DatabaseTools and DemoDatabaseTools. The new tool uses a JSON specification as input, so making changes to what needs to be added to the database no longer requires new code, only editing of a JSON file.

To support this new endeavor, a new SharedCommonCode object has been created, the SailPlatformServicesSession. This object is designed for reuse and will be integrated in the Orchestrator, the RootOfTrustProcess, or anything in C++ which makes API calls into the platform.

It should be noted that the RegisterDigitalContract (first step of the three step process of digital contract creation) has been modified so that it now returns the digital contract identifier of the new digital contract, along with the Status.

The new DatabaseInitializationTools can be used in three ways:

--step1 --> Registers organization, contacts, admins, users, dataset families, data federations
--step2 --> Registers datasets and digital contracts
--step3 --> Registers everything

By creating steps, it's now possible to initialize the database, then use the Dataset Tools to created datasets (separately), This prevents the chicken and end problem of the existing tool. However, if CSVP files already exist, then the tool can register everything all at once.