Rebased my local repo to fix a commit bug
There was two bugs. One is a commit issue whereas I made a change to my local repo which wasn't detected by the tool and wasn't uploaded (i.e. deleting the old code which called the demodatabasetool with the -d). The was previously committed, but it never got reflected in GitHub... Rebasing fixed the issue. The other bug is a , instead of a +