Bug fixes with Guid.h and Guid.cpp as well as packaging tools
New sample files have been created along with batch scripts to package tables as well as datasets. Errors were fixed in the Guid.h and Guid.cpp files in Linux and Windows and we are not forcing a eHyphensOnly default regardless of which of the old 'printing format' are used. Two new GUID print formats are made available, one is eForceRaw and the other is eForceHyphensAndCurlyBraces

Some improvements made to the DatabaseInitializationTool as well.